refactor(config): start splitting config.py
- Move config.py to config/config.py (will be split in the next commit)
- Move internal classes and functions to config/utils.py
- Make tests more independent of config module internals (use isolated_fs,
  import from core.constants instead of from core.config)
