+++ tests/cmd/auth/test_login.py
@@ -49,7 +49,7 @@ DT_FORMAT = "%Y-%m-%dT%H:%M:%S%z"
 @pytest.fixture(autouse=True)
 def tmp_config(monkeypatch, tmp_path):
     monkeypatch.setattr(
-        "ggshield.core.config.get_auth_config_dir", lambda: str(tmp_path)
+        "ggshield.core.config.utils.get_auth_config_dir", lambda: str(tmp_path)
     )
