+++ tests/cmd/test_config.py
@@ -40,7 +40,7 @@ expiry: not set
 @pytest.fixture(autouse=True)
 def tmp_config(monkeypatch, tmp_path):
     monkeypatch.setattr(
-        "ggshield.core.config.get_auth_config_dir", lambda: str(tmp_path)
+        "ggshield.core.config.utils.get_auth_config_dir", lambda: str(tmp_path)
     )
