+++ tests/cmd/test_ignore.py
@@ -1,7 +1,6 @@
 import os
 import tempfile
 
-from mock import patch
 from pygitguardian.models import Match, PolicyBreak
 
 from ggshield.cmd.secret.ignore import ignore_last_found
@@ -25,8 +24,6 @@ def compare_matches_ignore(match):
     return (match["name"], match["match"]) if isinstance(match, dict) else (match,)
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_cache_catches_last_found_secrets(client, isolated_fs):
     """
     GIVEN an empty cache and an empty config matches-ignore section
@@ -63,8 +60,6 @@ def test_cache_catches_last_found_secrets(client, isolated_fs):
     cache.load_cache()
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_cache_catches_nothing(client, isolated_fs):
     """
     GIVEN a cache of last found secrets same as config ignored-matches
@@ -92,8 +87,6 @@ def test_cache_catches_nothing(client, isolated_fs):
         assert cache.last_found_secrets == []
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_cache_old_config_no_new_secret(client, isolated_fs):
     """
     GIVEN a cache of last found secrets same as config ignored-matches
@@ -122,8 +115,6 @@ def test_cache_old_config_no_new_secret(client, isolated_fs):
         assert cache.last_found_secrets == []
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_ignore_last_found(client, isolated_fs):
     """
     GIVEN a cache of last found secrets not empty
@@ -145,8 +136,6 @@ def test_ignore_last_found(client, isolated_fs):
     assert cache.last_found_secrets == FOUND_SECRETS
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_ignore_last_found_with_manually_added_secrets(client, isolated_fs):
     """
     GIVEN a cache containing part of config ignored-matches secrets
@@ -169,8 +158,6 @@ def test_ignore_last_found_with_manually_added_secrets(client, isolated_fs):
     assert matches_ignore == found_secrets
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_do_not_duplicate_last_found_secrets(client, isolated_fs):
     """
     GIVEN 2 policy breaks on different files with the same ignore sha
@@ -188,8 +175,6 @@ def test_do_not_duplicate_last_found_secrets(client, isolated_fs):
     assert len(cache.last_found_secrets) == 1
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_do_not_add_policy_breaks_to_last_found(client, isolated_fs):
     """
     GIVEN 1 policy breaks on different files with the same ignore sha
@@ -206,8 +191,6 @@ def test_do_not_add_policy_breaks_to_last_found(client, isolated_fs):
     assert len(cache.last_found_secrets) == 0
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_ignore_last_found_preserve_previous_config(client, isolated_fs):
     """
     GIVEN a cache containing new secrets AND a config not empty
@@ -237,8 +220,6 @@ def test_ignore_last_found_preserve_previous_config(client, isolated_fs):
     assert config.exit_zero is True
 
 
-@patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [DOT_GITGUARDIAN_YAML])  # nosec
-@patch("ggshield.core.config.DEFAULT_LOCAL_CONFIG_PATH", DOT_GITGUARDIAN_YAML)  # nosec
 def test_ignore_last_found_compatible_with_previous_matches_ignore_format(
     client, isolated_fs
 ):
