+++ tests/core/test_cache.py
@@ -9,12 +9,12 @@ from ggshield.core.cache import Cache
 from ggshield.core.config import Config
 
 
+@pytest.mark.usefixtures("isolated_fs")
 class TestCache:
     def test_defaults(self, cli_fs_runner):
         cache = Cache()
         assert cache.last_found_secrets == []
 
-    @patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [".gitguardian.yml"])
     def test_load_cache_and_purge(self, cli_fs_runner):
         with open(".cache_ggshield", "w") as file:
             json.dump({"last_found_secrets": [{"name": "", "match": "XXX"}]}, file)
@@ -24,7 +24,6 @@ class TestCache:
         cache.purge()
         assert cache.last_found_secrets == []
 
-    @patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [".gitguardian.yml"])
     def test_load_invalid_cache(self, cli_fs_runner, capsys):
         with open(".cache_ggshield", "w") as file:
             json.dump({"invalid_option": True}, file)
@@ -33,7 +32,6 @@ class TestCache:
         captured = capsys.readouterr()
         assert "Unrecognized key in cache" in captured.out
 
-    @patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [".gitguardian.yml"])
     def test_save_cache(self, cli_fs_runner):
         with open(".cache_ggshield", "w") as file:
             json.dump({}, file)
@@ -61,7 +59,6 @@ class TestCache:
             open_mock.assert_called_once()
 
     @pytest.mark.parametrize("with_entry", [True, False])
-    @patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [".gitguardian.yml"])
     def test_save_cache_first_time(self, isolated_fs, with_entry):
         """
         GIVEN no existing cache
@@ -75,7 +72,6 @@ class TestCache:
 
         assert os.path.isfile(".cache_ggshield") is with_entry
 
-    @patch("ggshield.core.config.LOCAL_CONFIG_PATHS", [".gitguardian.yml"])
     def test_max_commits_for_hook_setting(self, cli_fs_runner):
         """
         GIVEN a yaml config with `max-commits-for-hook=75`
